Moved summary model
- Moved item summary model to summary namespace
